Use concat fragments instead of file resource templates for config
The net effect of this change is zero - it's backwards compatible
and does not result in any change for users.
The concat resources create the files with the same permissions
and the concat fragments take care of inserting the templates with
the same feature to override the template if need be.

However, since concat is now used, consumers of puppet-dns can now
insert their own concat fragments without having to maintain their
own template.
Before this change, if we wanted to diverge slightly from the
template, we would need to "fork" the template and maintain our own.
We could otherwise edit the file with resources like file_line but
this is not idempotent.

Spec test coverage for these bits have been improved a bit and
standardized as part of the commit.