Do not configure listen-on-v6 parameter if it's not set
The listen-on-v6 parameter would always be configured in
options.conf. This makes it unflexible with the use case of making
BIND listen on another port than 53.
Make it so the configuration is added only when the parameter is set
so that, if required, a user can set listen_on_v6 to false and add
his own configuration in additional_options.

Closes: #67