Change licence text in README to match LICENSE/metadata
The commit history shows the module has always been under the ASL-2.0,
and that the README text was an inconsistency in 999eb91 where the
LICENSE document was added and README.md was extended (probably from a
template).

Updated the main LICENSE document to include the full text, not only the
boilerplate copyright statement.

Fixes GH-66