Add manage_service parameter
This allows users to decide whether the module should manage the
dns service. The check defaults to true, but in case users want to
manage the service autonomously they can.
This also solves "duplicate declaration" issues when the service
is declared in another .pp file

This does not influence package installation.