add support for the statistics-channels directive
The statistics-channels statement declares communication channels
to be used by system administrators to get access to statistics
information of the name server. BIND will only start a listener
if the directive is found in the configuration. The ACLs are
rendered similar to the `controls` directive.

Statistics are not collected on a per-zone basis by default.
To enable this behaviour, the defined type for zones has a new
parameter to do so.